Lots of troubleshooting a problem that arose: The main thread hangs if it accesses sysfs too rapidly. I experimented with timings until I settled on a 50 ms delay before each and every sysfs access. That seems to let things work without the main thread hanging. The other thread continues to run, no problem, but I'm not able to kill the process once this happens.
Added BinarySempahore (a counting semaphore wasn't really appropriate
to manage the radio on/off state).

Added a delay after calling Radio::start(), to ensure the readio
got to its on/off wait before turning it on. This is a gross hack,
and if it proves to be necessary for the radio to be ready in this
way, there should be another semaphore back to the main thread. I
think, though, the delay can simply be removed, as it's not
necessary. I just though it might be the cause of the bad behavior
at first.

Some changes to the .dts file which are no longer needed now that
I'm using dtb-rebuilder.

Logging in GPIO can be removed. It was part of the debugging.